conversation going

+++ Job_Portail_Spring_Final/src/main/java/com/example/security/Other/ConversationParticipantId/ConversationParticipantId.java
@@ -10,6 +10,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 @EqualsAndHashCode
 public class ConversationParticipantId implements Serializable {
-    private Long conversation;
+    private Integer conversation;
     private Integer user;
 }
